passport_analysis_agent/pyproject.toml
[project]
name = "passport-analysis-agent"
version = "0.1.0"
description = "AI-powered passport document analysis agent using Claude Vision"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Passport Agent Team" }]
keywords = ["passport", "ocr", "mrz", "document-analysis", "claude", "ai"]

dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "pypdfium2>=4.30.0",
    "Pillow>=10.4.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "rapidfuzz>=3.0.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
passport-agent = "passport_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/passport_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
